chore: Move all components to 'src/components' folder and add JSDoc (as required by Styleguide)
