--- src/.eslintrc
@@ -1,5 +1,0 @@
-{
-  "env": {
-    "jest": true
-  }
-}
